requirement.txt
# Core
python-dotenv>=1.0.1
openai>=1.40.0
chromadb>=0.5.5

# Document loaders
pypdf>=4.2.0          # PDF parsing
python-docx>=1.1.0    # .docx support
beautifulsoup4>=4.12.3
markdown>=3.6
lxml>=5.2.1           # faster/more robust HTML parsing
pandas>=2.2.2         # for CSV/structured handling

# Retrieval (optional BM25 hybrid)
rank-bm25>=0.2.2

# API / Web
fastapi>=0.112.0
uvicorn[standard]>=0.30.0

# Utilities
tqdm>=4.66.4          # progress bars
